pyproject.toml
[project]
name = "loan-sanction-prediction"
version = "0.1.0"
description = "This is Loan Sanction Prediction project. One of my first and full fledged ML project."
authors = [{ name = "Althaf Muhammad" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "imbalanced-learn>=0.13.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "streamlit>=1.44.1",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "flash-lib>=0.1.1",
    "ipykernel>=6.29.5",
    "klib>=1.3.2",
    "loguru>=0.7.3",
    "matplotlib>=3.10.1",
    "plotly>=6.0.1",
    "pytest>=8.3.5",
    "ruff>=0.11.5",
    "seaborn>=0.13.2",
]
